+++ src/utils/parsingUtils.ts
@@ -9,9 +9,7 @@ export const isNumericString = (message: string) =>
     message.trim().length !== 0 && !Number.isNaN(Number(message))
 
 export const parseStringifiedDate = (value: any) => {
-    if (typeof value !== 'string') {
-        return null
-    }
+    if (typeof value !== 'string') return null
 
     // Remove quote from stringified date
     const temp = value.replace(/"/g, '')
@@ -23,9 +21,7 @@ export const parseStringifiedDate = (value: any) => {
     ) {
         const date = new Date(temp)
 
-        if (!Number.isNaN(date.getTime())) {
-            return date
-        }
+        if (!Number.isNaN(date.getTime())) return date
     }
 
     return null
@@ -50,27 +46,13 @@ export const parseStringifiedObject = (data: string) =>
     })
 
 export const parseStringifiedValue = (value: string) => {
-    if (!value) {
-        return value
-    }
-
-    if (isNumericString(value)) {
-        return +value
-    }
-
-    if (value === 'true') {
-        return true
-    }
-
-    if (value === 'false') {
-        return false
-    }
+    if (!value) return value
+    if (isNumericString(value)) return +value
+    if (value === 'true') return true
+    if (value === 'false') return false
 
     const date = parseStringifiedDate(value)
-
-    if (date) {
-        return date
-    }
+    if (date) return date
 
     if (isStringifiedObject(value)) {
         try {