+++ CHANGELOG.md
@@ -1,6 +1,28 @@
-# 1.3.3
+# 1.4.2 - 12 Oct 2025
+Bug fix:
+- [elysia#1466](https://github.com/elysiajs/elysia/issues/1466) / treaty2: unpack multiple SSE chunk
+
+# 1.4.1 - 14 Sep 2025
+Bug fix:
+- inline object value / Elysia file cause type error
+- macro should not mark property as required
+
+# 1.4.0 - 13 Sep 2025
+Improvement:
+- support Elysia 1.4
+
+# 1.3.3 - 24 Aug 2025
+Feature:
+- treaty2: support type safe Server-Sent Events (SSE)
+- treaty2: add utility type `Treaty.Data`, `Treaty.Error` to extract data and error type from a route
+
 Bug fix:
 - [elysia#823](https://github.com/elysiajs/elysia/issues/823) treaty2: not generating for dynamic params at root
+- treaty2: parse Date in object
+- treaty2: [#196](https://github.com/elysiajs/eden/issues/196) allow custom content-type
+
+Change:
+- minimum Elysia version is set to 1.3.18
 
 # 1.3.2 - 5 May 2025
 Bug fix:
@@ -16,7 +38,7 @@ Feature:
 - support Elysia 1.3
 
 Breaking Change:
-- [Treaty 2] drop the need for `.index()`
+- treaty2: drop the need for `.index()`
 
 # 1.2.0 - 23 Dec 2024
 Feature: